remove history from area
